Configure components to support multiple NFT standards
