+++ packages/desktop/views/dashboard/collectibles/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Irc27CollectibleDetails } from './Irc27CollectibleDetails.svelte'
